added temp and radio files

+++ Final Project/FinalProject/FinalProject/temperature.h
@@ -3,4 +3,4 @@
 */
 
 //Returns integer representation of the temperature
-int getTemp();
+uint8_t getTemp();
